Split settings Playwright flows
