+++ E2E/playwright/tests/harness-helpers.ts
@@ -13,3 +13,13 @@ export async function clickSidebarTool(page: Page, testID: string) {
   await openSidebarTools(page);
   await page.getByTestId(testID).click();
 }
+
+export async function openTopBarOverflow(page: Page) {
+  await page.getByTestId('top-bar-overflow-button').click();
+  await expect(page.getByTestId('top-bar-overflow-menu')).toHaveAttribute('open', '');
+}
+
+export async function openSettings(page: Page) {
+  await openTopBarOverflow(page);
+  await page.getByTestId('top-bar-overflow-settings').click();
+}
